+++ mono/zlib/adler32.c
