+++ mono/zlib/gzguts.h
