+++ mono/zlib/inffixed.h
