+++ support/minizip/ioapi.h
@@ -43,7 +43,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
-#include "zlib.h"
+#include "../mono/zlib/zlib.h"
 
 #if defined(USE_FILE32API)
 #define fopen64 fopen
