+++ support/minizip/unzip.c
@@ -72,7 +72,7 @@
         #define NOUNCRYPT
 #endif
 
-#include "zlib.h"
+#include "../mono/zlib/zlib.h"
 #include "unzip.h"
 
 #ifdef STDC
