+++ support/minizip/unzip.h
@@ -48,7 +48,7 @@ extern "C" {
 #endif
 
 #ifndef _ZLIB_H
-#include "zlib.h"
+#include "../mono/zlib/zlib.h"
 #endif
 
 #ifndef  _ZLIBIOAPI_H
