+++ support/minizip/zip.c
@@ -26,7 +26,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
-#include "zlib.h"
+#include "../mono/zlib/zlib.h"
 #include "zip.h"
 
 #ifdef STDC
